requirements.txt
python-binance
python-telegram-bot==21.9
python-dotenv
openai
